Refactor: Rename and move Book Interface to Contract's folder
