+++ src/app/Providers/RepositoryServiceProvider.php
@@ -3,7 +3,7 @@
 namespace App\Providers;
 
 use Illuminate\Support\ServiceProvider;
-use App\Http\Repositories\IBookRepository;
+use App\Http\Repositories\Contracts\BookContract;
 use App\Http\Repositories\BookRepository;
 
 class RepositoryServiceProvider extends ServiceProvider
@@ -15,7 +15,7 @@ class RepositoryServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        $this->app->bind(IBookRepository::class, BookRepository::class);
+        $this->app->bind(BookContract::class, BookRepository::class);
     }
 
     /**
